backend/requirements.txt
Django>=5,<6
djangorestframework
django-cors-headers
pillow
numpy
opencv-python
dj-database-url
psycopg2-binary
gunicorn
whitenoise
python-dotenv
inference-sdk==0.51.4
dj-database-url
whitenoise>=6.6
